rag-chatbot-poc/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.15
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
python-pptx==0.6.23
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
pyyaml==6.0.1
psutil==5.9.6